[AArch64, ILP32] Add change to gcc.target/aarch64/vect_smlal_1.c that was missed from the previous commit (r201166).
From-SVN: r201175

+++ gcc/testsuite/gcc.target/aarch64/vect_smlal_1.c
@@ -12,7 +12,7 @@ typedef signed long long *__restrict__ pS64_t;
 typedef unsigned char U8_t;
 typedef unsigned short U16_t;
 typedef unsigned int U32_t;
-typedef unsigned long U64_t;
+typedef unsigned long long U64_t;
 typedef unsigned char *__restrict__ pU8_t;
 typedef unsigned short *__restrict__ pU16_t;
 typedef unsigned int *__restrict__ pU32_t;
